+++ app/controllers/bands_controller.rb
