+++ app/controllers/cities_controller.rb
@@ -1,9 +1,5 @@
-module Api
-  module V1
-    class CitiesController < ApplicationController
+class CitiesController < ApplicationController
       def index
         render json: City.includes(:users), include: ('users')
       end
-    end
   end
-end
