+++ app/controllers/friend_requests_controller.rb
