+++ app/controllers/genres_controller.rb
