+++ app/serializers/city_serializer.rb
@@ -1,4 +1,3 @@
 class CitySerializer < ActiveModel::Serializer
   attributes :id, :name
-  has_many :users
 end
