+++ app/serializers/user_serializer.rb
@@ -1,4 +1,4 @@
 class UserSerializer < ActiveModel::Serializer
-  attributes :id, :name, :city, :instruments, :genres, :conversations, :messages, :friend_requests,:networks
+  attributes :id, :name
 end
