+++ config/initializers/active_model_serializer.rb
@@ -1,12 +1,12 @@
 
-ActiveModelSerializers.config.adapter = :json_api
+# ActiveModelSerializers.config.adapter = :json_api
 
-api_mime_types = %W(
- application/vnd.api+json
- text/x-json
- application/json
-)
-Mime::Type.register 'application/vnd.api+json', :json, api_mime_types
+# api_mime_types = %W(
+#  application/vnd.api+json
+#  text/x-json
+#  application/json
+# )
+# Mime::Type.register 'application/vnd.api+json', :json, api_mime_types
 
 ActiveModelSerializers.config.default_includes = '**'
